+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [feattle 4.0.0] - 2025-05-07
+
+### Changed
+
+- `axum_router()` takes a second argument with the logic to extract the name of the user that modified a feattle
+
+## [feattle-ui 4.0.0] - 2025-05-07
+
+### Changed
+
+- `axum_router()` takes a second argument with the logic to extract the name of the user that modified a feattle
+
 ## [feattle-ui 3.0.1] - 2025-04-17
 
 ### Fixed